+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -5,8 +5,8 @@ import com.techcourse.model.User;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import org.apache.catalina.controller.AbstractController;
-import org.apache.catalina.controller.util.StaticResourceLoader;
+import org.apache.catalina.handler.AbstractController;
+import org.apache.catalina.handler.util.StaticResourceLoader;
 import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
