+++ tomcat/src/main/java/com/techcourse/controller/StaticResourceController.java
@@ -3,9 +3,9 @@ package com.techcourse.controller;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import org.apache.catalina.controller.AbstractController;
-import org.apache.catalina.controller.util.ContentTypeExtractor;
-import org.apache.catalina.controller.util.StaticResourceLoader;
+import org.apache.catalina.handler.AbstractController;
+import org.apache.catalina.handler.util.ContentTypeExtractor;
+import org.apache.catalina.handler.util.StaticResourceLoader;
 import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
