+++ tomcat/src/main/java/org/apache/catalina/handler/util/ContentTypeExtractor.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.controller.util;
+package org.apache.catalina.handler.util;
 
 import java.util.Map;
