+++ tomcat/src/main/java/org/apache/catalina/handler/util/StaticResourceLoader.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.controller.util;
+package org.apache.catalina.handler.util;
 
 import java.io.IOException;
 import java.io.InputStream;
