+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -3,8 +3,8 @@ package org.apache.coyote.http11;
 import com.techcourse.exception.UncheckedServletException;
 import java.io.IOException;
 import java.net.Socket;
-import org.apache.catalina.controller.Controller;
-import org.apache.catalina.controller.RequestMapping;
+import org.apache.catalina.handler.Controller;
+import org.apache.catalina.handler.RequestMapping;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.HttpRequestBuilder;
